# openapi.yaml
openapi: 3.0.3
info:
  title: Secure Legal DMS API
  version: 0.1.0
  description: >
    Backend contract for the SIH 2026 secure document management system.
    This file is the source of truth. Do not change an endpoint without
    telling the whole team in writing.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - sessionToken: []
paths:
...
  /cases:
    get:
      tags: [cases]
      summary: List cases the caller is assigned to.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Assigned cases only, never all cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  total:
                    type: integer
    post:
      tags: [cases]
      summary: >
        Open a case (case.create). Station and creator come from the
        session; the creator is assigned in the same transaction.
        Audited as case_create.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [case_number, title]
              properties:
                case_number:
                  type: string
                title:
                  type: string
                sensitivity:
                  $ref: '#/components/schemas/Sensitivity'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Case number already exists
...
  /documents:
    post:
      tags: [documents]
      summary: >
        File evidence. Creates version 1. The database assigns a
        type-specific evidence number within the case (FIR-001, STM-002).
        Audited as upload in the same transaction.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [case_id, file]
              properties:
                case_id:
                  type: string
                  format: uuid
                title:
                  type: string
                  description: Optional description; defaults to the evidence number
                doc_type:
                  $ref: '#/components/schemas/DocType'
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: >
            Stored and hashed. Ledger anchoring and OCR run asynchronously,
            so anchor_status will usually be "pending" here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: File too large
...
  /icjs/fir:
    get:
      tags: [icjs]
      summary: >
        FIR reference lookup against ICJS. Fixture-backed mock.
        The FIR number is a query parameter, not a path segment, because
        FIR numbers contain slashes and an encoded %2F in a path is
        normalised or rejected by most reverse proxies.
      parameters:
        - name: fir_number
          in: query
          required: true
          schema:
            type: string
          example: FIR/0142/2026
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IcjsFir'
        '400':
          description: fir_number is required
        '404':
          $ref: '#/components/responses/NotFound'
  /icjs/cases/{case_id}/link:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [icjs]
      summary: >
        Link a case to an FIR held in CCTNS. Returns the acknowledgement
        ICJS would return; persists nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fir_number]
              properties:
                fir_number:
                  type: string
      responses:
        '200':
          description: Linked
        '400':
          description: fir_number is required
        '404':
          $ref: '#/components/responses/NotFound'
  /icjs/documents/{document_id}/payload:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [icjs]
      summary: >
        The exact metadata payload that would be pushed to ICJS for this
        document. Read-only; transmits nothing. Carries a pointer and a
        hash, never the file, its text, or any victim data.
      responses:
        '200':
          description: OK
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Opaque session token from /auth/mfa/verify
  parameters:
    CaseId:
      name: case_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
  schemas:
    DocType:
      type: string
      enum: [fir, statement, forensic_report, charge_sheet, court_filing, notice, other]
    Rank:
      type: string
      enum: [constable, head_constable, sub_inspector, inspector, dcp, prosecutor, judge, forensic_analyst]
    Sensitivity:
      type: string
      description: >
        "protected" triggers mandatory redaction on every export path,
        for offences covered by Section 72 BNS.
      enum: [normal, restricted, protected]
...
    IcjsFir:
      type: object
      description: FIR reference held in CCTNS. Fixture-backed mock.
      properties:
        fir_number:
          type: string
          example: FIR/0142/2026
        cctns_id:
          type: string
        police_station:
          type: string
        district:
          type: string
        state:
          type: string
        registered_on:
          type: string
          format: date-time
        act:
          type: string
        sections:
          type: array
          items:
            type: string
        investigating_officer:
          type: object
          properties:
            name:
              type: string
            service_number:
              type: string
            rank:
              $ref: '#/components/schemas/Rank'
        status:
          type: string
        accused_count:
          type: integer
        source:
          type: string
        mock:
          type: boolean
          description: Always true until a real ICJS connection exists.
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        service_number:
          type: string
        rank:
          $ref: '#/components/schemas/Rank'
        station:
          type: string
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        case_number:
          type: string
          example: FIR/0142/2026
        title:
          type: string
        sensitivity:
          $ref: '#/components/schemas/Sensitivity'
        status:
          type: string
          enum: [open, under_investigation, charge_sheeted, closed]
        created_at:
          type: string
          format: date-time
...
    DocumentVersion:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        version:
          type: integer
        sha256:
          type: string
          description: Hash of the original bytes, computed before encryption
          example: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
        size_bytes:
          type: integer
        uploaded_by:
          $ref: '#/components/schemas/User'
        uploaded_at:
          type: string
          format: date-time
        anchor_status:
          type: string
          enum: [pending, anchored, failed]
        ledger_tx_id:
          type: string
          nullable: true
        ocr_status:
          type: string
          enum: [pending, done, failed]
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: forbidden
        message:
          type: string
          example: You are not assigned to this case.
  responses:
...
    Forbidden:
      description: >
        Logged in but not permitted. Always logged as access_denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: >
        Returned instead of 403 where confirming existence would itself
        leak information about a protected case.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'